fixed the register service
