+++ server/src/routes/loginRegisterRoutes.js
@@ -1,0 +1,9 @@
+const express = require('express');
+const router = express.Router();
+
+const loginRegisterController = require('../controllers/loginRegisterController'); 
+
+router.post('/register', loginRegisterController.registerController); 
+router.post('/login', loginRegisterController.loginController);
+
+module.exports = router;
